pyproject.toml
[tool.poetry]
name = "cloner"
version = "0.1.0"
description = ""
authors = ["Alan Bohannon <alanbohannon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
